ChatWithPDF/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
entrypoints==0.4
faiss-cpu==1.13.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
ollama==0.6.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==3.20.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
semver==3.0.4
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
streamlit==1.18.1
tenacity==9.1.2
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
validators==0.35.0
watchdog==6.0.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
